Rename 'brakepoints' to 'breakpoints' for consistency
Renamed files and package names from 'brakepoints' to 'breakpoints' to correct a spelling error. Updated all relevant import statements in GdScriptDebugAdapterSupportProvider.kt accordingly.
