+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/godot/run/configurations/gdscript/breakpoints/GdScriptExceptionBreakpointType.kt
@@ -1,4 +1,4 @@
-package com.jetbrains.rider.plugins.godot.run.configurations.gdscript.brakepoints
+package com.jetbrains.rider.plugins.godot.run.configurations.gdscript.breakpoints
 
 import com.intellij.xdebugger.breakpoints.XBreakpoint
 import com.intellij.xdebugger.breakpoints.XBreakpointProperties
